+++ example/ext/sqla/db/versions/1162611363e2_create_user_table.py
@@ -20,7 +20,7 @@ def upgrade():
     op.create_table(
         'project',
         sa.Column('id', sa.Integer, primary_key=True),
-        sa.Column('name', sa.String(50), nullable=False),
+        sa.Column('name', sa.String(50), nullable=False, unique=True),
     )