+++ example/ext/sqla/resources/__init__.py
@@ -1,14 +1,7 @@
-import logging
-
-
-from ..model import Session
-
 from royal.ext.sqla import Collection as SACollection
 from royal.ext.sqla import Item as SAItem
 
-log = logging.getLogger(__name__)
-
-from pyramid.tweens import EXCVIEW
+from ..model import Session
 
 
 class CollectionBase(SACollection):
@@ -20,20 +13,4 @@ class ItemBase(SAItem):
 
 
 def includeme(config):
-
     config.add_resource('projects')
-    config.add_tween('example.ext.sqla.resources.session_tween_factory',
-                     under=EXCVIEW)
-    config.scan(__name__)
-
-
-def remove_session(request):
-    Session.remove()
-
-
-def session_tween_factory(handler, registry):
-    def session_tween(request):
-        request.add_finished_callback(remove_session)
-        response = handler(request)
-        return response
-    return session_tween